refactor(targets): move notify net helpers from utils
